In order to maintain compatibility with older trees, we now have
minadbd.old and minui.old. I had to use a TARGET_GLOBAL_CFLAG to
handle ifdef issues in minui/minui.d because healthd includes
minui/minui.h and there was no other alternative to make minui.h
compatible with older trees without having to modify healthd rules
which is outside of TWRP.

Note that the new minui does not currently have support for qcom
overlay graphics. Support for this graphics mode will likely be
added in a later patch set. If you are building in a 6.0 tree and
have a device that needs qcom overlay graphics, be warned, as off
mode charging may not work properly. A dead battery in this case
could potentially brick your device if it is unable to charge as
healthd handles charging duties.

Update rules for building toolbox and add rules for making toybox

Use permissive.sh in init.rc which will follow symlinks so we do
not have to worry about what binary is supplying the setenforce
functionality (toolbox, toybox, or busybox).

Fix a few warnings in the main recovery binary source code.

Fix a few includes that were missing that prevented compiling in
6.0

+++ tarWrite.c
@@ -17,6 +17,9 @@
 */
 
 #include <fcntl.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
 #include "libtar/libtar.h"
 #include "twcommon.h"
 
@@ -71,7 +74,7 @@ ssize_t write_libtar_buffer(int fd, const void *buffer, size_t size) {
 			// nothing to write
 			return 0;
 		}
-		if (write(fd, write_buffer, buffer_loc) != buffer_loc) {
+		if (write(fd, write_buffer, buffer_loc) != (int)buffer_loc) {
 			LOGERR("Error writing tar file!\n");
 			buffer_loc = 0;
 			return -1;
